requirements.txt
aiohttp==3.7.4.post0
alabaster==0.7.12
APScheduler==3.7.0
apturl==0.5.2
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dateparser==1.0.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.17.1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
httplib2==0.14.0
idna==2.8
imagesize==1.2.0
Jinja2==3.0.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
monotonic==1.5
multidict==5.1.0
netifaces==0.10.4
notify-run==0.0.14
numpy==1.21.1
oauthlib==3.1.0
olefile==0.46
packaging==21.0
pandas==1.3.0
paramiko==2.6.0
pbr==5.6.0
pdoc3==0.9.2
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
Pygments==2.9.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
PyQRCode==1.2.1
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-binance==1.0.12
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
regex==2021.7.6
requests==2.22.0
requests-unixsocket==0.2.0
roman==2.0.0
schedule==1.1.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
snowballstemmer==2.1.0
Sphinx==4.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
systemd-python==234
testresources==2.0.1
typing-extensions==3.10.0.0
tzlocal==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==4.0.2
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
websockets==9.1
xkit==0.0.0
yarl==1.6.3
